Implement player score and respective communication between server and client.
Implement SMSG_DISCONNECT to inform client that the server was shutdown.
Refactor to the Message class and respective subclasses in order to take the switch statements from the HandleMessage Method in Message class. Now the HandleMessage calls the Handle method from the message class that calls the respective message handler method in the handler class.
Move ServerHandler and ClientMessage to the bombermen project as it became cycle because of last refactor.